FIX: Handling of test result codes in case of assert errors in multithreading
Removes false output of 'invalid return code'

This is a follow-up fix for an earlier fix. When issuing an assert we terminate the running thread, leaving the 'return code' from the test-function in an undefined state.

The return code is now defined (for debugging) - this will check if we have asserts/errors which would have caused the termination and therefore leading to a missing return code.