NEW: Testing different ways to provide extensions.
This is using a 'Query' COM-like method. By adding a single function (QueryInterface) in the ITesting - interface we can request access to other parts of the system.

This also implements a small 'Config' interface which can list parts of the test-runners configuration. Not sure that is needed - never had the need...

This also changes the ITesting pre/post and allows them to return kTR_xxx