NEW: Version handling some what implemented
The idea seems to work. The public available testinterface.h defines a weak symbol with the version. This is checked in the test-runner which then selects the correct interface before calling.

In essence you can combine testable libraries of different versions in one run - AS LONG AS each library is compiled against the same version..