NEW: Exposing process-exit status from 'waitpid'
I just had the return code from spawn previously exposed. But if a test in a module will crash the whole trun child process - I wouldn't detect it. This will detect the crash and print 'abnormal termination' for that module.

This is not yet exposed in the JSON report (module will simply not be reported at all).