FIX: In case of V1 we switch to the 'kThreadedWithExit' by default
V1 interfaces don't have return values for assert errors. Thus we can not signal HOW to proceed back to the function under test. In Version1 the default was to terminate the thread running the test.

Before execution of a test-library we check the library version - in case it is using V1 we switch execution type to 'threaded_with_exit' - this allows 'Asserts' (and others) to terminate the test from within the test-runner.